[FIX] 컴포넌트 구성 변경
