+++ src/components/Navs.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
 //react-bootstrap
 import Container from 'react-bootstrap/Container';
