--- src/components/dashboard/sidebar/SidebarIcon/Activity.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-const IconActivity = () => {
-  return (
-    <svg
-      width='20'
-      height='20'
-      viewBox='0 0 22 22'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M11 0C17.0751 0 22 4.92487 22 11C22 17.0751 17.0751 22 11 22C4.92487 22 0 17.0751 0 11C0 4.92487 4.92487 0 11 0ZM11 2C6.02944 2 2 6.02944 2 11C2 15.9706 6.02944 20 11 20C15.9706 20 20 15.9706 20 11C20 6.02944 15.9706 2 11 2ZM11 4C11.5523 4 12 4.44772 12 5V10.5109L16.1139 13.7106C16.5499 14.0497 16.6284 14.678 16.2894 15.1139C15.9503 15.5499 15.322 15.6284 14.8861 15.2894L10.3861 11.7894C10.1425 11.5999 10 11.3086 10 11V5C10 4.44772 10.4477 4 11 4Z'
-      />
-    </svg>
-  )
-}
-
-export default IconActivity
