--- src/components/dashboard/sidebar/SidebarIcon/Schedule.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-const IconSchedule = () => {
-  return (
-    <svg
-      width='21'
-      height='20'
-      viewBox='0 0 22 21'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M16 2V1C16 0.447715 16.4477 0 17 0C17.5523 0 18 0.447715 18 1V2H19C20.6569 2 22 3.34315 22 5V8V10V15C22 18.3137 19.3137 21 16 21H6C2.68629 21 0 18.3137 0 15V10V8V5C0 3.34315 1.34315 2 3 2H5V1C5 0.447715 5.44772 0 6 0C6.55228 0 7 0.447715 7 1V2H16ZM2 8H20V5C20 4.44772 19.5523 4 19 4H18C18 4.55228 17.5523 5 17 5C16.4477 5 16 4.55228 16 4H7C7 4.55228 6.55228 5 6 5C5.44772 5 5 4.55228 5 4H3C2.44772 4 2 4.44772 2 5V8ZM2 10V15C2 17.2091 3.79086 19 6 19H16C18.2091 19 20 17.2091 20 15V10H2Z'
-      />
-    </svg>
-  )
-}
-
-export default IconSchedule
