--- src/components/dashboard/sidebar/SidebarIcon/Settings.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-const IconSettings = () => {
-  return (
-    <svg
-      width='20'
-      height='22'
-      viewBox='0 0 20 22'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M7.09429 2.33907C7.38357 0.97584 8.59581 0 10 0C11.4042 0 12.6164 0.97584 12.9057 2.33907L12.962 2.60457C14.0187 2.97101 14.9989 3.53018 15.8497 4.2566L16.1074 4.17298C17.442 3.73991 18.9 4.29353 19.6021 5.5C20.3042 6.70647 20.0585 8.23593 19.0131 9.16609L18.8113 9.3457C19.02 10.4399 19.018 11.5617 18.8095 12.6528L19.0131 12.8339C20.0585 13.7641 20.3042 15.2935 19.6021 16.5C18.9 17.7065 17.442 18.2601 16.1074 17.827L15.8489 17.7431C15.4276 18.1021 14.971 18.4234 14.4827 18.7031C13.9945 18.9827 13.4857 19.2144 12.9617 19.3969L12.9057 19.6609C12.6164 21.0242 11.4042 22 10 22C8.59581 22 7.38357 21.0242 7.09429 19.6609L7.03795 19.3954C5.98129 19.029 5.00107 18.4698 4.1503 17.7434L3.8926 17.827C2.55798 18.2601 1.10004 17.7065 0.397941 16.5C-0.304156 15.2935 -0.05845 13.7641 0.986893 12.8339L1.18874 12.6543C0.980025 11.5601 0.982019 10.4383 1.19048 9.34725L0.986893 9.16609C-0.05845 8.23593 -0.304156 6.70647 0.397941 5.5C1.10004 4.29353 2.55798 3.73991 3.8926 4.17298L4.1511 4.25686C4.57236 3.89792 5.02899 3.5766 5.51725 3.29693C6.00551 3.01726 6.51431 2.78559 7.03826 2.60311L7.09429 2.33907ZM10 2C9.5491 2 9.15985 2.31335 9.06696 2.75109L8.88929 3.58834C8.81147 3.95506 8.53351 4.24784 8.16917 4.34682C7.59911 4.50169 7.04777 4.72966 6.52521 5.02898C6.00265 5.3283 5.52798 5.68802 5.10776 6.10038C4.83919 6.36392 4.44464 6.45636 4.0856 6.33986L3.26593 6.07388C2.83737 5.93482 2.36922 6.1126 2.14377 6.5C1.91832 6.8874 1.99722 7.37852 2.33289 7.6772L2.97653 8.24993C3.25789 8.50029 3.37444 8.88578 3.27841 9.24845C2.97456 10.396 2.97134 11.6029 3.27656 12.7524C3.37291 13.1152 3.25641 13.501 2.97486 13.7516L2.33289 14.3228C1.99722 14.6215 1.91832 15.1126 2.14377 15.5C2.36922 15.8874 2.83737 16.0652 3.26593 15.9261L4.08555 15.6602C4.44502 15.5435 4.84005 15.6363 5.10861 15.9005C5.95939 16.7375 7.01454 17.3382 8.16816 17.6509C8.53275 17.7497 8.81098 18.0426 8.88884 18.4095L9.06696 19.2489C9.15985 19.6867 9.5491 20 10 20C10.4509 20 10.8402 19.6867 10.933 19.2489L11.1107 18.4117C11.1885 18.0449 11.4665 17.7522 11.8308 17.6532C12.4009 17.4983 12.9522 17.2703 13.4748 16.971C13.9974 16.6717 14.472 16.312 14.8922 15.8996C15.1608 15.6361 15.5554 15.5436 15.9144 15.6601L16.7341 15.9261C17.1626 16.0652 17.6308 15.8874 17.8562 15.5C18.0817 15.1126 18.0028 14.6215 17.6671 14.3228L17.0235 13.7501C16.7421 13.4997 16.6256 13.1142 16.7216 12.7516C17.0254 11.604 17.0287 10.3971 16.7234 9.24763C16.6271 8.88479 16.7436 8.49897 17.0251 8.24844L17.6671 7.6772C18.0028 7.37852 18.0817 6.8874 17.8562 6.5C17.6308 6.1126 17.1626 5.93482 16.7341 6.07388L15.9144 6.33984C15.555 6.45649 15.1599 6.36367 14.8914 6.09947C14.0406 5.2625 12.9855 4.6618 11.8318 4.3491C11.4672 4.25027 11.189 3.95739 11.1112 3.59046L10.933 2.75109C10.8402 2.31335 10.4509 2 10 2ZM10 14.6316C12.0216 14.6316 13.6605 13.0057 13.6605 11C13.6605 8.99433 12.0216 7.36842 10 7.36842C7.97838 7.36842 6.33953 8.99433 6.33953 11C6.33953 13.0057 7.97838 14.6316 10 14.6316ZM10 12.6316C9.09174 12.6316 8.35544 11.9011 8.35544 11C8.35544 10.0989 9.09174 9.36842 10 9.36842C10.9083 9.36842 11.6446 10.0989 11.6446 11C11.6446 11.9011 10.9083 12.6316 10 12.6316Z'
-      />
-    </svg>
-  )
-}
-
-export default IconSettings
