--- src/components/dashboard/sidebar/SidebarIcon/Tasks.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-const IconTasks = () => {
-  return (
-    <svg
-      width='20'
-      height='20'
-      viewBox='0 0 22 22'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M7 0H15C18.866 0 22 3.13401 22 7V15C22 18.866 18.866 22 15 22H7C3.13401 22 0 18.866 0 15V7C0 3.13401 3.13401 0 7 0ZM7 2C4.23858 2 2 4.23858 2 7V15C2 17.7614 4.23858 20 7 20H15C17.7614 20 20 17.7614 20 15V7C20 4.23858 17.7614 2 15 2H7ZM7.69164 10.2784L9.96275 12.5203L14.2178 7.36355C14.5693 6.93756 15.1996 6.87718 15.6256 7.22868C16.0516 7.58018 16.1119 8.21046 15.7604 8.63645L10.8096 14.6364C10.4357 15.0895 9.75378 15.1243 9.33576 14.7117L6.28661 11.7018C5.89356 11.3138 5.88946 10.6806 6.27745 10.2876C6.66544 9.89453 7.29859 9.89043 7.69164 10.2784Z'
-      />
-    </svg>
-  )
-}
-
-export default IconTasks
